+++ src/main/java/com/development/order/model/entities/Product.java
@@ -20,7 +20,7 @@ public class Product implements Serializable{
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	
 	private String name;
 	private Double price;
@@ -37,7 +37,7 @@ public class Product implements Serializable{
 		
 	}
 	
-	public Product(Integer id,String name, Double price, String type, Seller seller) {
+	public Product(Long id,String name, Double price, String type, Seller seller) {
 		this.id = id;
 		this.name = name;
 		this.price = price;
@@ -50,11 +50,11 @@ public class Product implements Serializable{
 		return pkgs;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
