+++ src/main/java/com/development/order/repositories/ShippingRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.development.order.model.entities.Shipping;
 
-public interface ShippingRepository extends JpaRepository<Shipping, Integer> {
+public interface ShippingRepository extends JpaRepository<Shipping, Long> {
 
 }
